refactor: 펫시터 정보 확인 ts로 변환
