+++ .prettierrc
@@ -1,3 +1,3 @@
 {
-  "endOfLine": "auto"
+  "endOfLine": "lf"
 }
